pyproject.toml
[project]
name = "track"
version = "0.1.0"
readme = "README.md"
authors = [{ name = "Stanley5249", email = "seer852741@gmail.com" }]
requires-python = ">=3.12"
dynamic = ["description"]
dependencies = ["sugar-cli>=0.0.1.dev0"]

[build-system]
requires = ["flit_core>=3.2,<4"]
build-backend = "flit_core.buildapi"
